+++ docs/changelog.md
@@ -18,6 +18,7 @@ Sprint 7 / February 29, 2020
 - updated Timer and Button styles.
 - removed Date Component file and combined Timeline component and Date component into a single compound component.
 - routed up Header logo to be clickable Link that navigates to the Dashboard page.
+- added Alignment component.
 
 ## Release 0.6.0
 Sprint 6 / February 15, 2020